requirements.txt
Flask
gunicorn
sklearn
numpy
pandas
matplotlib
PyYAML==5.1
evidently
dill
-e .